Move manifest helpers to utils
